// lib/types/subjectProperty.ts
export interface SubjectProperty {
  address: string
  zip: string
  lat: number
  lng: number
city: string
state: string
  squareFeet: number
  yearBuilt: number
  beds: number
  baths: number
  propertyType: "SFR"
  hasGarage: boolean
  isBusyStreet: boolean
  nearCommercial: boolean
  nearDuplex: boolean
  nearLargeApartment: boolean
  nearFreeway: boolean
}